Removing Bot as this looks very odd
Bot got tagged at top of the page and does not get visible appropriately. hence removed the same from the page.
Again we needed proper container to add in HTML to host this bot and request for authentication for this bot as well.

Does not serve the purpose for me at this point in time. Hence discarding the functionality.